+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 0.1.6 / 2016-09-01
 ------------------
-- Handling promises from the fetch api.
+- Processing of fetch promises.
 
 0.1.5 / 2016-08-30
 ------------------